moved main code for user creation to own module.
Tests are not included because they were written for
the handler and the class UserEntriesCreatorForPerson has no tests of its own.
