+++ user-creation/build.gradle
@@ -6,13 +6,13 @@ plugins {
 
 dependencies {
 
+
     implementation project(":json")
     implementation project(":dynamo-client")
-    implementation project(":customer-commons")
-    implementation project(":user-access-commons")
-    implementation project(":user-access-service")
-    implementation project(":user-access-public-model")
-    implementation project(":user-creation")
+    implementation project (":customer-commons")
+    implementation project (":user-access-commons")
+    implementation project (":user-access-service")
+    implementation project (":user-access-public-model")
 
     implementation libs.nva.apigatewayv2
     implementation libs.aws.java.sdk2.cognitoidp
@@ -24,20 +24,22 @@ dependencies {
     implementation libs.nva.auth
     implementation libs.aws.lambda.events
 
+
     implementation libs.aws.sdk2.dynamodbenhanced
     implementation libs.aws.sdk2.dynamodb
     implementation libs.aws.sdk2.urlconnection
     implementation libs.aws.sdk2.secrets
 
+
     testImplementation libs.slf4j
     testImplementation libs.guava
     testImplementation libs.nva.events
-    testImplementation project(":user-access-internal-model")
-    testImplementation project(":customer-testing")
+    testImplementation project (":user-access-internal-model")
+    testImplementation project (":customer-testing")
     testImplementation project(':user-access-testing')
 }
 
-test {
+test{
     environment "COGNITO_HOST", "example.com"
     environment "AWS_REGION", "eu-west-1"
     environment "API_DOMAIN", "example.com"
