+++ user-creation/src/main/java/no/unit/nva/useraccessservice/usercreation/cristin/person/CristinAffiliation.java
@@ -1,4 +1,4 @@
-package no.unit.nva.cognito.cristin.person;
+package no.unit.nva.useraccessservice.usercreation.cristin.person;
 
 import static nva.commons.core.attempt.Try.attempt;
 import com.fasterxml.jackson.annotation.JsonIgnore;
